requirements.txt
psycopg2==2.9.10
pyodbc==5.2.0
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
python-socketio==5.10.0
eventlet==0.33.3
python-dotenv==1.0.0
jsonschema==4.20.0
tabulate==0.9.0

